Add fake metadata store and fix tests. (#958)
* Add fake metadata store and fix tests.

Also, add instructions on how to build/run the backend with Bazel.

Note that the fake metadata store works, but I need to add proper tests
that exercise it. That'll be done in a separate PR.

One thing I'm missing here is how to make Bazel run well in Travis. I
will send a follow up PR for doing this.

* move select for update to the db interface