feat(backend): Support pipeline parameters macros in one-off runs (#5086)
* [Backend] Move ParameterFormatter to common/util

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* [Backend] Make resource mngr use params formatter for simple Runs

This commit integrates the workflow resource manager with the
ParameterFormatter, to replace two macros: ``[[RunUUID]]`` (which
replaces with the run UUID) and ``[[CurrentTime.<template>]]`` which
replaces with the time when the workflow is submitted to the
API Server.

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* [Backend] Validate param pointer is not null

Pipelines can be created with parameters without default values. If such
a pipeline is run without providing a value to these parameters that
don't have defaults, `param.Value` will be `nil`

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* Move -1 placeholder to constant

Signed-off-by: Stefano Fioravanzo <stefano@arrikto.com>

* fixup! [Backend] Make resource mngr use params formatter for simple Runs